fix: extract range segments in mixin class
